\name{boxByConfidence}
\alias{boxByConfidence}
\title{Plotting performance: Box plots for target value confidence-coverage}
\usage{
  boxByConfidence(res)
}
\arguments{
  \item{res}{The collected results from a computer
  experiment.}
}
\value{
  A plot object.
}
\description{
  Plotting performance: Box plots for target value
  confidence-coverage
}

