\name{boxByMethod}
\alias{boxByMethod}
\title{Plotting performance: Box plots for target value confidence-coverage}
\usage{
  boxByMethod(res)
}
\arguments{
  \item{res}{The collected results from a computer
  experiment.}
}
\value{
  A plot object.
}
\description{
  Plotting performance: Box plots for target value
  confidence-coverage
}

