\name{collectAllExperiments}
\alias{collectAllExperiments}
\title{Running a computer experiment -- Collect all the results}
\usage{
  collectAllExperiments(reg)
}
\arguments{
  \item{reg}{A valid registry generated by
  'makeExperimentRegistry'.}
}
\value{
  List of data frames containing the performance measures
  of all point and interval estimates for the heterogeneity
  and the regression coefficients.
}
\description{
  Collects all the results of all finished experiments in
  the given registry for all predefined algorithms.
}

