\name{collectExperiments}
\alias{collectExperiments}
\title{Running a computer experiment in batch mode}
\usage{
  collectExperiments(reg, pattern)
}
\arguments{
  \item{reg}{A valid registry generated by
  'makeExperimentRegistry'.}

  \item{pattern}{string containing the algorithm pattern
  for which the collection shall be performed.}
}
\value{
  List of data frame containing the performance measures of
  all point and interval estimates for the heterogeneity
  and the regression coefficients.
}
\description{
  Collects the results of all finished experiments in the
  given registry for a given pattern.
}

