\name{lenBoxByMethod}
\alias{lenBoxByMethod}
\title{Plotting performance: Box plot of mean width}
\usage{
  lenBoxByMethod(res)
}
\arguments{
  \item{res}{: The collected results from a computer
  experiment.}
}
\value{
  A plot object.
}
\description{
  Plotting performance: Box plot of mean width
}

