\name{performanceConfH}
\alias{performanceConfH}
\title{Running a computer experiment: Adding performance measures}
\usage{
  performanceConfH(accum_int, true)
}
\arguments{
  \item{accum_int}{accumulated interval estimates.  At
  least the following columns need to be present: lower and
  upper.}

  \item{true}{true parameter.}
}
\description{
  Adding performance measurements to accumulated results of
  a computer experiment running multiple analysis of
  different simulated data following a random effects meta
  regression model.
}
\details{
  Adds performance measurements to interval estimates of
  the heterogeneity.
}
\examples{
# For an example, see the 'performance' function.
}

