\name{performancePointH}
\alias{performancePointH}
\title{Running a computer experiment: Adding performance measures}
\usage{
  performancePointH(point, h)
}
\arguments{
  \item{point}{accumulated point estimates.}

  \item{h}{true parameter.}
}
\description{
  Adding performance measurements to accumulated results of
  a computer experiment running multiple analysis of
  different simulated data following a random effects meta
  regression model.
}
\details{
  Adds performance measurements to point estimates of the
  heterogeneity.
}
\examples{
# For an example, see the 'performance' function.
}

