\name{performancePointR}
\alias{performancePointR}
\title{Running a computer experiment: Adding performance measures}
\usage{
  performancePointR(point, b)
}
\arguments{
  \item{point}{accumulated point estimates.}

  \item{b}{true parameter.}
}
\description{
  Adding performance measurements to accumulated results of
  a computer experiment running multiple analysis of
  different simulated data following a random effects meta
  regression model.
}
\details{
  Adds performance measurements to point estimates of the
  regression coefficients.
}
\examples{
# For an example, see the 'performance' function.
}

