\name{plotCoefficientInterval}
\alias{plotCoefficientInterval}
\title{Plot pivots: Interval estimates of the heterogeneity}
\usage{
  plotCoefficientInterval(cnfr)
}
\arguments{
  \item{cnfr}{interval estimates of the heterogeneity.}
}
\description{
  Plot pivots: Interval estimates of the heterogeneity
}

