\name{plotHeterogeneityInterval}
\alias{plotHeterogeneityInterval}
\title{Plot pivots: Interval estimates of the heterogeneity}
\usage{
  plotHeterogeneityInterval(cnfh)
}
\arguments{
  \item{cnfh}{interval estimates of the heterogeneity.}
}
\description{
  Plot pivots: Interval estimates of the heterogeneity
}

