\name{plotIntervalEstimates}
\alias{plotIntervalEstimates}
\title{Example: Plotting interval estimates}
\usage{
  plotIntervalEstimates(cnf)
}
\arguments{
  \item{cnf}{data frame of interval estimates}
}
\value{
  An object created by ggplot2.
}
\description{
  Plots a graphical representation of interval estimates in
  the data frame 'cnf' by type of method used for the
  estimation.
}

