\name{plotStudyUnbalance}
\alias{plotStudyUnbalance}
\title{Example: Plotting study unbalances in group assignments}
\usage{
  plotStudyUnbalance(dat)
}
\arguments{
  \item{dat}{data frame of study responses of binomial
  type.}
}
\value{
  An object created by ggplot2.
}
\description{
  Example: Plotting study unbalances in group assignments
}

