\name{render}
\alias{render}
\title{Render plot: To PDF}
\usage{
  render(name, plotObj, path, scale = 1, height = 11.6,
    width = 8.2)
}
\arguments{
  \item{name}{Self explenatory}

  \item{plotObj}{Self explenatory}

  \item{path}{Self explenatory}

  \item{scale}{Self explenatory}

  \item{height}{Self explenatory}

  \item{width}{Self explenatory}
}
\description{
  Renders obj into a pdf-file of name: path++name.
}

