\name{renderSVG}
\alias{renderSVG}
\title{Render plot: To SVG}
\usage{
  renderSVG(name, plotObj, path, scale = 1, height = 11.6,
    width = 8.2)
}
\arguments{
  \item{name}{Should be self explanatory.}

  \item{plotObj}{Should be self explanatory.}

  \item{path}{Should be self explanatory.}

  \item{scale}{Should be self explanatory.}

  \item{height}{Should be self explanatory.}

  \item{width}{Should be self explanatory.}
}
\description{
  Renders obj into a svg-file of name: path++name.  Neat
  feature is that the default size in A4.  Simply use the
  `scale` parameter to adjust the size of the plot to a
  fraction of a page.
}

