\name{sctBias}
\alias{sctBias}
\title{Plotting performance: Scatter plots against heterogeneity}
\usage{
  sctBias(res, ...)
}
\arguments{
  \item{res}{The collected results from a computer
  experiment.}

  \item{...}{further argumets to scale_y_continuous}
}
\value{
  A plot object.
}
\description{
  Plotting performance: Scatter plots against heterogeneity
}

