\name{sctVersusH}
\alias{sctVersusH}
\title{Plotting performance: Scatter plot against heterogeneity}
\usage{
  sctVersusH(res)
}
\arguments{
  \item{res}{: The collected interval results from a
  computer experiment.}
}
\value{
  A plot object.
}
\description{
  Plotting performance: Scatter plot against heterogeneity
}

