\name{sdmByMethod}
\alias{sdmByMethod}
\title{Plotting performance: Scatter plot against heterogeneity}
\usage{
  sdmByMethod(res)
}
\arguments{
  \item{res}{The collected results from a computer
  experiment.}
}
\value{
  A plot object.
}
\description{
  Plotting performance: Scatter plot against heterogeneity
}

