\name{sdmByType}
\alias{sdmByType}
\title{Plotting performance: Scatter plot against heterogeneity}
\usage{
  sdmByType(res)
}
\arguments{
  \item{res}{: The collected results from a computer
  experiment.}
}
\value{
  A plot object.
}
\description{
  Plotting performance: Scatter plot against heterogeneity
}

