\name{sdsByType}
\alias{sdsByType}
\title{Plotting performance: Scatter plot against heteroscedasticity}
\usage{
  sdsByType(res)
}
\arguments{
  \item{res}{The collected results from a computer
  experiment.}
}
\value{
  A plot object.
}
\description{
  Plotting performance: Scatter plot against
  heteroscedasticity
}

