\name{sctMSE}
\alias{sctMSE}
\title{Plotting performance: Scatter plots against heterogeneity}
\usage{
  sctMSE(res, ...)
}
\arguments{
  \item{res}{The collected results from a computer
  experiment.}

  \item{...}{further argumets to scale_y_continuous}
}
\value{
  A plot object.
}
\description{
  Plotting performance: Scatter plots against heterogeneity
}

